+++ app/controllers/wishlists_controller.ts
@@ -4,23 +4,27 @@ import { WishlistStoreValidator } from '#validators/wishlist_validator'
 import type { HttpContext } from '@adonisjs/core/http'
 
 export default class WishlistsController {
-    async index({view}:HttpContext){
-        const userId=1
-        const wishlists=await Wishlist.query().where('userId', userId).preload('product')
+    async index({view, auth}:HttpContext){
+        const user=auth.user!
+        const wishlists=await Wishlist.query().where('userId', user.id).preload('product')
         return view.render('pages/wishlist', {wishlists})
     }
 
-    async store({request, response}: HttpContext){
+    async store({request, response, auth}: HttpContext){
         const data =await request.validateUsing(WishlistStoreValidator)
-        const userId=1
+        const user=auth.user!
         await Product.findOrFail(data.productId)
-        await Wishlist.firstOrCreate({userId, productId:data.productId})
-        return response.redirect('/wishlists')
+        await Wishlist.firstOrCreate({userId:user.id, productId:data.productId})
+        return response.redirect('/wishlist')
     }
 
-    async destroy({params, response}:HttpContext){
+    async destroy({params, response, auth}:HttpContext){
+        const user=auth.user!
         const item=await Wishlist.findOrFail(params.id)
+        if (item.userId !== user.id){
+            return response.unauthorized('Not Allowed')
+        }
         await item.delete()
-        return response.redirect('/wishlists')
+        return response.redirect('/wishlist')
     }
 }
